+++ src/cpp11.cpp
@@ -5,21 +5,21 @@
 #include "cpp11/declarations.hpp"
 #include <R_ext/Visibility.h>
 
-// shikokuchuo.cpp
+// windowfn.cpp
 cpp11::doubles maxOver(const cpp11::doubles& x, int window);
 extern "C" SEXP _ichimoku_maxOver(SEXP x, SEXP window) {
   BEGIN_CPP11
     return cpp11::as_sexp(maxOver(cpp11::as_cpp<cpp11::decay_t<const cpp11::doubles&>>(x), cpp11::as_cpp<cpp11::decay_t<int>>(window)));
   END_CPP11
 }
-// shikokuchuo.cpp
+// windowfn.cpp
 cpp11::doubles minOver(const cpp11::doubles& x, int window);
 extern "C" SEXP _ichimoku_minOver(SEXP x, SEXP window) {
   BEGIN_CPP11
     return cpp11::as_sexp(minOver(cpp11::as_cpp<cpp11::decay_t<const cpp11::doubles&>>(x), cpp11::as_cpp<cpp11::decay_t<int>>(window)));
   END_CPP11
 }
-// shikokuchuo.cpp
+// windowfn.cpp
 cpp11::doubles meanOver(const cpp11::doubles& x, int window);
 extern "C" SEXP _ichimoku_meanOver(SEXP x, SEXP window) {
   BEGIN_CPP11